work on ToDatexExpressionData in derive macro
