+++ crates/datex-core/src/ast/expressions.rs
@@ -370,47 +370,11 @@ pub struct BinaryOperation {
     pub right: DatexExpression,
     pub ty: Option<Type>,
 }
-impl ToInstructions for BinaryOperation {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            yield RegularInstruction::binary_operation(self.operator);
-            for instruction in self.left.to_instructions(ctx) {
-                yield instruction;
-            }
-            for instruction in self.right.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
-}
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct RangeDeclaration {
     pub start: DatexExpression,
     pub end: DatexExpression,
-}
-impl ToInstructions for RangeDeclaration {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            yield RegularInstruction::range();
-            for instruction in self.start.to_instructions(ctx) {
-                yield instruction;
-            }
-            for instruction in self.end.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -420,63 +384,11 @@ pub struct ComparisonOperation {
     pub right: DatexExpression,
 }
 
-impl ToInstructions for ComparisonOperation {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            yield RegularInstruction::comparison_operation(self.operator);
-            for instruction in self.left.to_instructions(ctx) {
-                yield instruction;
-            }
-            for instruction in self.right.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
-}
-
 #[derive(Clone, Debug, PartialEq)]
 pub struct UnboxAssignment {
     pub operator: Option<ModificationOperator>,
     pub unbox_expression: DatexExpression,
     pub assigned_expression: DatexExpression,
-}
-impl ToInstructions for UnboxAssignment {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            match self.operator {
-                Some(operator) => match operator {
-                    ModificationOperator::AddAssign => {
-                        yield RegularInstruction::increment()
-                    }
-                    ModificationOperator::SubtractAssign => {
-                        yield RegularInstruction::decrement()
-                    }
-                    _ => todo!("Generate x = x * z instructions;"),
-                },
-                None => yield RegularInstruction::set_shared_container_value(),
-            };
-
-            // compile assigned expression
-            for instruction in self.assigned_expression.to_instructions(ctx) {
-                yield instruction;
-            }
-
-            // compile unbox expression
-            for instruction in self.unbox_expression.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -492,54 +404,6 @@ pub struct PropertyAssignment {
     pub base: DatexExpression,
     pub property: DatexExpression,
     pub assigned_expression: DatexExpression,
-}
-
-impl ToInstructions for PropertyAssignment {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            let PropertyAssignment {
-                base,
-                property,
-                assigned_expression,
-                ..
-            } = self;
-            // depending on the key, handle different property assignments
-
-            match &property.data() {
-                DatexExpressionData::Text(key)
-                    if key.0.len() <= u8::MAX as usize =>
-                {
-                    yield RegularInstruction::set_entry_text(key.0.clone());
-                }
-
-                DatexExpressionData::Integer(index)
-                    if let Some(index) = index.as_u32() =>
-                {
-                    yield RegularInstruction::set_entry_index(index);
-                }
-
-                _ => {
-                    for instruction in property.to_instructions(ctx) {
-                        yield instruction;
-                    }
-                }
-            }
-            // compile assigned expression
-            for instruction in assigned_expression.to_instructions(ctx) {
-                yield instruction;
-            }
-
-            // compile base expression
-            for instruction in base.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -561,21 +425,6 @@ pub struct TypeDeclarationExpression {
 pub struct UnaryOperation {
     pub operator: UnaryOperator,
     pub expression: DatexExpression,
-}
-impl ToInstructions for UnaryOperation {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            yield RegularInstruction::unary_operation(self.operator);
-            for instruction in self.expression.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
 }
 
 #[derive(Clone, Debug, PartialEq, Default, Hash, Eq)]
@@ -607,89 +456,11 @@ pub struct Apply {
     pub arguments: Vec<DatexExpression>,
 }
 
-impl ToInstructions for Apply {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            yield RegularInstruction::apply(self.arguments.len() as u8);
-            // compile arguments
-            for argument in &self.arguments {
-                for instruction in argument.to_instructions(ctx) {
-                    yield instruction;
-                }
-            }
-            // compile function expression
-            for instruction in self.base.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
-}
-
 #[derive(Clone, Debug, PartialEq)]
 pub struct InterfaceMethodCall {
     pub target: DatexExpression,
     pub method_name: String,
     pub arguments: Vec<DatexExpression>,
-}
-impl ToInstructions for InterfaceMethodCall {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            // TODO: replace with trait impls
-            match self.method_name.as_str() {
-                "append" => {
-                    yield RegularInstruction::append_entry();
-
-                    for argument in &self.arguments {
-                        for instruction in argument.to_instructions(ctx) {
-                            yield instruction;
-                        }
-                    }
-                }
-
-                "clear" => {
-                    yield RegularInstruction::clear();
-                }
-
-                "splice" => {
-                    yield RegularInstruction::splice_dynamic();
-
-                    for argument in &self.arguments {
-                        for instruction in argument.to_instructions(ctx) {
-                            yield instruction;
-                        }
-                    }
-                }
-
-                _ => {
-                    yield RegularInstruction::call_method(
-                        self.method_name.clone(),
-                        self.arguments.len() as u8,
-                    );
-
-                    for argument in &self.arguments {
-                        for instruction in argument.to_instructions(ctx) {
-                            yield instruction;
-                        }
-                    }
-                }
-            }
-
-            // compile target expression
-            for instruction in self.target.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
 }
 
 impl InterfaceMethodCall {
@@ -712,59 +483,10 @@ pub struct PropertyAccess {
     pub property: DatexExpression,
 }
 
-impl ToInstructions for PropertyAccess {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            // depending on the key, handle different property accesses
-            match self.property.data() {
-                // simple text key if length fits in u8
-                DatexExpressionData::Text(key) if key.0.len() <= 255 => {
-                    yield RegularInstruction::get_entry_text(key.0.clone());
-                }
-                // index access if integer fits in u32
-                DatexExpressionData::Integer(index)
-                    if let Some(index) = index.as_u32() =>
-                {
-                    yield RegularInstruction::get_entry_index(index);
-                }
-                _ => {
-                    yield RegularInstruction::get_entry_dynamic();
-                    for instruction in self.property.to_instructions(ctx) {
-                        yield instruction;
-                    }
-                }
-            }
-
-            // compile base expression
-            for instruction in self.base.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
-}
-
 #[derive(Clone, Debug, PartialEq)]
 pub struct GenericInstantiation {
     pub base: DatexExpression,
     pub generic_arguments: Vec<TypeExpression>,
-}
-impl ToInstructions for GenericInstantiation {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            // NOTE: might already be handled in type compilation
-            todo!()
-        })
-    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -854,19 +576,7 @@ pub struct RequestSharedRef {
     pub mutability: ReferenceMutability,
     pub address: PointerAddress,
 }
-impl ToInstructions for RequestSharedRef {
-    type InstructionType = RegularInstruction;
 
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(core::iter::once(RegularInstruction::get_shared_ref(
-            self.address.clone(),
-            &self.mutability,
-        )))
-    }
-}
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct CallableDeclaration {
@@ -889,23 +599,6 @@ pub struct CallableSignature {
 #[derive(Clone, Debug, PartialEq)]
 pub struct List {
     pub items: Vec<DatexExpression>,
-}
-impl ToInstructions for List {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            yield RegularInstruction::list(self.items.len() as u32);
-            for item in &self.items {
-                for instruction in item.to_instructions(ctx) {
-                    yield instruction;
-                }
-            }
-        })
-    }
 }
 
 impl List {
@@ -930,46 +623,6 @@ impl FromIterator<(DatexExpression, DatexExpression)> for Map {
     fn from_iter<T: IntoIterator<Item = (DatexExpression, DatexExpression)>>(iter: T) -> Self {
         let entries = iter.into_iter().collect();
         Map { entries }
-    }
-}
-
-impl ToInstructions for Map {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            yield RegularInstruction::map(self.entries.len() as u32);
-            for (key, value) in &self.entries {
-                match &*key.data {
-                    // text -> insert key string
-                    DatexExpressionData::Text(text) => {
-                        if text.len() < 256 {
-                            yield RegularInstruction::key_value_short_text(
-                                text.0.clone(),
-                            );
-                        } else {
-                            yield RegularInstruction::key_value_dynamic();
-                            yield RegularInstruction::text(text.0.clone());
-                        }
-                    }
-                    // other -> insert key as dynamic
-                    _ => {
-                        yield RegularInstruction::key_value_dynamic();
-                        for instruction in key.to_instructions(ctx) {
-                            yield instruction;
-                        }
-                    }
-                };
-
-                // value
-                for instruction in value.to_instructions(ctx) {
-                    yield instruction;
-                }
-            }
-        })
     }
 }
 
@@ -1000,26 +653,6 @@ pub struct TagExpression {
     pub expression: Option<DatexExpression>,
 }
 
-impl ToInstructions for TagExpression {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            yield RegularInstruction::tagged_value(
-                self.tag.clone(),
-                self.expression.is_none(),
-            );
-            if let Some(expression) = &self.expression {
-                for instruction in expression.to_instructions(ctx) {
-                    yield instruction;
-                }
-            }
-        })
-    }
-}
 
 impl Display for TagExpression {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
@@ -1032,20 +665,6 @@ pub struct RootPropertyAccess {
     pub property_name: String,
 }
 
-impl ToInstructions for RootPropertyAccess {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            let root_property = RootProperty::from_str(&self.property_name)
-                .expect("invalid root property name");
-            yield RegularInstruction::get_root_property(root_property);
-        })
-    }
-}
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct StackAssignment {
@@ -1077,21 +696,6 @@ pub struct Unbox {
     pub expression: DatexExpression,
 }
 
-impl ToInstructions for Unbox {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            yield RegularInstruction::unbox();
-            for instruction in self.expression.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
-}
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct CloneExpression {
@@ -1104,50 +708,12 @@ pub struct DeriveRef {
     pub expression: DatexExpression,
 }
 
-impl ToInstructions for DeriveRef {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            for instruction in self.expression.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
-}
-
 #[derive(Clone, Debug, PartialEq)]
 pub struct CreateShared {
     pub mutability: SharedContainerMutability,
     pub expression: DatexExpression,
 }
 
-impl ToInstructions for CreateShared {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            match self.mutability {
-                SharedContainerMutability::Immutable => {
-                    yield RegularInstruction::create_shared();
-                }
-                SharedContainerMutability::Mutable => {
-                    yield RegularInstruction::create_shared_mut();
-                }
-            }
-
-            for instruction in self.expression.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
-}
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct DeriveSharedRef {
@@ -1155,29 +721,6 @@ pub struct DeriveSharedRef {
     pub expression: DatexExpression,
 }
 
-impl ToInstructions for DeriveSharedRef {
-    type InstructionType = RegularInstruction;
-
-    fn to_instructions<'tracking, 'ctx, 'iter>(
-        &'iter self,
-        ctx: &'iter InstructionContext<'tracking, 'ctx>,
-    ) -> Box<impl Iterator<Item = Self::InstructionType> + 'iter> {
-        Box::new(gen move {
-            match self.mutability {
-                ReferenceMutability::Immutable => {
-                    yield RegularInstruction::derive_shared_reference();
-                }
-                ReferenceMutability::Mutable => {
-                    yield RegularInstruction::derive_shared_reference_mut();
-                }
-            }
-
-            for instruction in self.expression.to_instructions(ctx) {
-                yield instruction;
-            }
-        })
-    }
-}
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct CreateMut {
