+++ crates/datex-core/src/shared_values/owned_shared_container/mod.rs
@@ -1,6 +1,8 @@
 mod clone_unsafe;
 mod common;
 pub mod datex_proxy;
+#[cfg(feature = "decompiler")]
+mod to_datex_expression_data;
 
 use crate::{
     prelude::*,
