+++ crates/datex-core/src/shared_values/owned_shared_container/to_datex_expression_data.rs
@@ -1,0 +1,18 @@
+use crate::ast::expressions::{CreateShared, DatexExpressionData};
+use crate::ast::spanned::Spanned;
+use crate::shared_values::OwnedSharedContainer;
+use crate::shared_values::traits::SharedContainerCommon;
+use crate::traits::to_datex_expression_data::ToDatexExpressionData;
+
+impl ToDatexExpressionData for OwnedSharedContainer {
+    fn to_datex_expression_data(&self) -> DatexExpressionData {
+        if self.is_borrowed() {
+            DatexExpressionData::OmitRecursive
+        } else {
+            DatexExpressionData::CreateShared(CreateShared {
+                mutability: self.container_mutability(),
+                expression: (&*self.value_container()).to_datex_expression_data().with_default_span(),
+            })
+        }
+    }
+}
