+++ crates/datex-core/src/shared_values/referenced_shared_container/to_datex_expression_data.rs
@@ -1,9 +1,10 @@
 use crate::ast::expressions::{CreateShared, DatexExpressionData};
 use crate::ast::spanned::Spanned;
+use crate::shared_values::ReferencedSharedContainer;
 use crate::shared_values::traits::SharedContainerCommon;
 use crate::traits::to_datex_expression_data::ToDatexExpressionData;
 
-impl<T: SharedContainerCommon> ToDatexExpressionData for T {
+impl ToDatexExpressionData for ReferencedSharedContainer {
     fn to_datex_expression_data(&self) -> DatexExpressionData {
         if self.is_borrowed() {
             DatexExpressionData::OmitRecursive
