+++ crates/datex-core/src/values/cast/rust_values/duration.rs
@@ -1,13 +1,14 @@
 #[cfg(feature = "decompiler")]
 mod to_datex_expression_data {
+    use core::time::Duration;
     use crate::ast::expressions::DatexExpressionData;
     use crate::traits::to_datex_expression_data::ToDatexExpressionData;
-    use crate::values::core_values::text::Text;
-    use crate::prelude::*;
+    use crate::values::core_values::integer::Integer;
 
-    impl ToDatexExpressionData for String {
+    impl ToDatexExpressionData for Duration {
         fn to_datex_expression_data(&self) -> DatexExpressionData {
-            DatexExpressionData::Text(Text(self.clone()))
+            // TODO: use amount once implemented
+            DatexExpressionData::Integer(Integer::from(self.as_millis()))
         }
     }
 }
